Refactor Most Failed Tests widget
