--- app/src/pages/inside/dashboardPage/widgets/failedTests/index.js
@@ -1,3 +1,0 @@
-import FailedTests from './failedTests';
-
-export { FailedTests };
