+++ app/src/pages/inside/dashboardPage/widgets/mostFailedTests/index.js
@@ -1,0 +1,3 @@
+import MostFailedTests from './mostFailedTests';
+
+export { MostFailedTests };
