+++ app/src/pages/inside/dashboardPage/widgets/mostFailedTests/mostFailedTestsTable/index.jsx
@@ -2,10 +2,10 @@ import * as React from 'react';
 import { func } from 'prop-types';
 import classNames from 'classnames/bind';
 import { ScrollWrapper } from 'components/main/scrollWrapper';
-import FailedTestsTableHeader from './failedTestsTableHeader';
-import FailedTestsTableBody from './failedTestsTableBody';
-import { PTTests } from './pTypes';
-import styles from './failedTests.scss';
+import FailedTestsTableHeader from './mostFailedTestsTableHeader';
+import FailedTestsTableBody from './mostFailedTestsTableBody';
+import { PTTests } from '../pTypes';
+import styles from '../mostFailedTests.scss';
 
 const cx = classNames.bind(styles);
